Add files via upload
Once we have created some of the engine, we can consolidate it to one file. It is separated or now to allow us to pick a couple files to start on and build from there.

These files are simple "Hello World" scripts, they won't work off the bat either because they aren't named "main" anymore lol. Either way, just open and edit the code to calculate/create functions/values so that we can use them in the engine.